requirements.txt
# Core dependencies
requests>=2.32.3
python-dotenv>=1.0.0
pyperclip>=1.11.0

# File watching
watchdog>=6.0.0

# Scheduling
schedule>=1.2.2

# Gmail API
google-auth>=2.48.0
google-auth-oauthlib>=1.2.4
google-auth-httplib2>=0.2.1
google-api-python-client>=2.187.0

# WhatsApp automation
playwright>=1.58.0

# Testing
pytest>=9.0.2
pytest-asyncio>=1.3.0
